chore(lrm/server): migrate utils to ts
